CU-8693xmupv: Front end changes for project groups
